+++ src/Louis.Hosting/Internal/EventIds.cs
@@ -9,18 +9,19 @@ internal static class EventIds
     {
         public const int StateChanged = 1;
         public const int BeforeSetup = 2;
-        public const int SetupCompleted = 3;
-        public const int SetupCanceled = 4;
-        public const int SetupFailed = 5;
-        public const int BeforeExecute = 6;
-        public const int ExecuteCompleted = 7;
-        public const int ExecuteCanceled = 8;
-        public const int ExecuteFailed = 9;
-        public const int BeforeTeardown = 10;
-        public const int TeardownCompleted = 11;
-        public const int TeardownFailed = 12;
-        public const int StopRequested = 13;
-        public const int HostedServiceStarting = 14;
-        public const int HostedServiceStopping = 15;
+        public const int SetupSuccessful = 3;
+        public const int SetupNotSuccessful = 4;
+        public const int SetupCanceled = 5;
+        public const int SetupFailed = 6;
+        public const int BeforeExecute = 7;
+        public const int ExecuteCompleted = 8;
+        public const int ExecuteCanceled = 9;
+        public const int ExecuteFailed = 10;
+        public const int BeforeTeardown = 11;
+        public const int TeardownCompleted = 12;
+        public const int TeardownFailed = 13;
+        public const int StopRequested = 14;
+        public const int HostedServiceStarting = 15;
+        public const int HostedServiceStopping = 16;
     }
 }